+++ src/comfyui_queue_manager/queue_manager.py
@@ -1,5 +1,4 @@
 # Add custom API routes, using router
-from .qm_gallery import QM_Gallery
 from .qm_options import QM_Options
 
 from .qm_queue import QM_Queue
@@ -12,7 +11,6 @@ class QueueManager:
         init_schema()
         self.options = QM_Options()
         self.queue = QM_Queue(self)
-        self.gallery = QM_Gallery(self)
         self.server = QM_Server(self, __version__)
 
         return